@@ -4,49 +4,54 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "start:zh-hk": "ng serve --configuration=zh-hk --port=4201",
+    "start:zh-cn": "ng serve --configuration=zh-cn --port=4202",
     "build": "ng build",
+    "build:i18n": "bash i18n_build.sh",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "int:extract": "ng xi18n --output-path src/locale"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.7",
-    "@angular/cdk": "^8.2.0",
-    "@angular/common": "~8.2.3",
-    "@angular/compiler": "~8.2.3",
-    "@angular/core": "~8.2.3",
-    "@angular/forms": "~8.2.3",
-    "@angular/material": "^8.2.0",
+    "@angular/animations": "^9.0.0-rc.6",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~9.0.0-rc.6",
+    "@angular/compiler": "~9.0.0-rc.6",
+    "@angular/core": "~9.0.0-rc.6",
+    "@angular/forms": "~9.0.0-rc.6",
+    "@angular/localize": "~9.0.0-rc.6",
+    "@angular/material": "^8.2.3",
     "@angular/material-moment-adapter": "^8.2.3",
-    "@angular/platform-browser": "~8.2.3",
-    "@angular/platform-browser-dynamic": "~8.2.3",
-    "@angular/router": "~8.2.3",
-    "marked": "^0.7.0",
+    "@angular/platform-browser": "~9.0.0-rc.6",
+    "@angular/platform-browser-dynamic": "~9.0.0-rc.6",
+    "@angular/router": "~9.0.0-rc.6",
+    "marked": "^0.8.0",
     "moment": "^2.24.0",
     "rxjs": "~6.5.3",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.1",
-    "@angular/cli": "~8.3.1",
-    "@angular/compiler-cli": "~8.2.3",
-    "@angular/language-service": "~8.2.3",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.900.0-rc.6",
+    "@angular/cli": "~9.0.0-rc.6",
+    "@angular/compiler-cli": "~9.0.0-rc.6",
+    "@angular/language-service": "~9.0.0-rc.6",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "3.6.4"
   }
 }
