@@ -6,15 +6,22 @@ import { UserAuthContext } from '../models/UserAuthContext';
 
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
-    constructor(private auth: AuthenticationService) { }
+    protected authContext: UserAuthContext;
+
+    constructor(private authService: AuthenticationService) {
+        this.authService.getAuthContextStream()
+            .subscribe((authContext: UserAuthContext) => {
+                this.authContext = authContext;
+            });
+
+    }
 
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         // add authorization header with jwt token if available
-        const authContext: UserAuthContext = this.auth.getAuthContext();
-        if (authContext && authContext.token) {
+        if (this.authContext && this.authContext.token) {
             request = request.clone({
                 setHeaders: {
-                    Authorization: `Bearer ${authContext.token}`
+                    Authorization: `Bearer ${this.authContext.token}`
                 }
             });
         }
