@@ -9,6 +9,8 @@ import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/c
 import { LoadingService } from 'src/app/services/loading.service';
 import { ReservationService } from 'src/app/services/reservation.service';
 import { to12HourClock } from 'src/app/utils/reservationDisplay';
+import { LocaleService } from 'src/app/services/locale.service';
+import { TranslationService } from 'src/app/services/translationService';
 
 @Component({
   selector: 'app-reservation-management',
@@ -20,7 +22,9 @@ export class ReservationManagementComponent implements OnInit {
 
   constructor(private reservationService: ReservationService,
     private loadingService: LoadingService,
-    public dialog: MatDialog) { }
+    public dialog: MatDialog,
+    private localeService: LocaleService,
+    private translationService: TranslationService) { }
 
   ngOnInit() {
     this.getReservations();
@@ -36,8 +40,8 @@ export class ReservationManagementComponent implements OnInit {
   onCancel(reservation: ReservationView) {
     const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
       data: {
-        title: 'Cancel Reservation',
-        text: 'Are you sure you want to cancel this reservation?',
+        title: $localize `:@@modal.cancelReservation.title:Cancel Reservation`,
+        text: $localize `:@@modal.cancelReservation.text:Are you sure you want to cancel this reservation?`,
         yesNo: true
       },
       width: DIALOG_WIDTHS.CONFIRMATION
@@ -51,17 +55,17 @@ export class ReservationManagementComponent implements OnInit {
   }
   
   private cancelReservation(reservation: ReservationView) {
+    reservation.desc = this.translationService.translateRole(reservation.desc);
     const cancelObs = this.reservationService.cancelReservation(reservation.id)
       .pipe(tap(() => { this.getReservations(); }));
 
     this.loadingService.callWithLoader(cancelObs, [
-      { state: LoadState.Loading, title: 'Cancelling', text: 'Sending Cancellation ...' },
-      { state: LoadState.Complete, title: 'Cancelled', text: [
-        `We have cancelled your session for  `,
-        `**${reservation.desc}** on  `,
-        `**${formatDate(reservation.startDate, 'fullDate', 'en-US')}** at **${to12HourClock(reservation.startTime)}**.`
-      ].join('\n') },
-      { state: LoadState.Error, title: 'Cancellation Error' }
+      { state: LoadState.Loading, title: $localize `:@@mySchedule.cancel.loader.title:Cancelling`, text: $localize `:@@mySchedule.cancel.loader.text:Sending Cancellation ...` },
+      { state: LoadState.Complete, title: $localize `:@@mySchedule.cancel.loader.done.title:Cancelled`, text: $localize
+`:@@mySchedule.cancel.loader.done.text:We have cancelled your session for  
+**${reservation.desc}** on  
+**${formatDate(reservation.startDate, 'fullDate', this.localeService.locale)}** at **${to12HourClock(reservation.startTime)}**.`},
+      { state: LoadState.Error, title: $localize `:@@mySchedule.cancel.loader.error:Cancellation Error` }
     ]);
   }
 }
