@@ -6,7 +6,9 @@ import { EmailValidationCredentials } from 'src/app/models/EmailValidationCreden
 import { LoginCredentials } from 'src/app/models/LoginCredentials';
 import { UserAuthContext } from 'src/app/models/UserAuthContext';
 import { environment } from 'src/environments/environment';
-import { errorsAreFalse, map200toTrue } from '../utils/httpUtil';
+import { errorsAreFalse, map200toTrue } from 'src/app/utils/httpUtil';
+import * as docCookies from 'src/app/utils/docCookies';
+import { SESSION_AUTH_CONTEXT, COOKIE_AUTH_CONTEXT } from 'src/app/constants/storage-constants';
 
 const API_URL = environment.API_URL;
 
@@ -16,39 +18,43 @@ const API_URL = environment.API_URL;
 export class AuthenticationService {
   private authContextEmitter: Subject<UserAuthContext> = new ReplaySubject(1);
   private $authContext: Observable<UserAuthContext> = this.authContextEmitter.pipe(shareReplay(1));
+  private refreshTimeout: any = undefined;
 
-  constructor(private http: HttpClient) { }
-
-  getSecurityClearance(): number {
-    const authContext = this.getAuthContext();
-    return authContext ? authContext.securityClearance : 0;
+  constructor(private http: HttpClient) {
+    // set up first replay authContext value
+    const initialAuthContext = this.getAuthContext();
+    this.authContextEmitter.next(initialAuthContext);
+    if (initialAuthContext && initialAuthContext.securityClearance) {
+      this.setupRefreshTimeout();
+    }
   }
 
   login(username: string, password: string): Observable<string> {
     const url = `${API_URL}/login`;
     const cred = new LoginCredentials(username, password);
+    this.refreshTimeout = undefined;
 
     if (username && username.length > 0 &&
       password && password.length > 0) {
       return this.http.post<any>(url, cred)
         .pipe(
           map((response: any) => {
-            const authContext = new UserAuthContext(username, response.token, response.securityClearance, response.name);
-            sessionStorage.setItem('currentUser', JSON.stringify(authContext));
+            const authContext = new UserAuthContext(username, response.token, response.securityClearance, response.name, response.expiry);
+            sessionStorage.setItem(SESSION_AUTH_CONTEXT, JSON.stringify(authContext));
 
             // emit for subject
             this.authContextEmitter.next(authContext);
-
-            // refresh token 30 seconds before expiry
-            const expiryTime = response.expiry - Date.now();
-            setTimeout(() => { this.refresh(authContext) }, expiryTime - 30000);
+            this.setupRefreshTimeout();
             return null;
           }),
           catchError((error: any) => {
-            if (error.status === 401) {
-              return of('Invalid Login Credentials');
-            } else if (error.status === 400) {
-              return of('Email is missing or unvalidated.');
+            if (error.status === 400) {
+              if (error.error.error === "Invalid login credentials") {
+                return of($localize `:@@error.login.credential:Invalid Login Credentials.`);
+              }
+              else {
+                return of($localize `:@@error.login.invalidEmail:Email is missing or unvalidated.`);
+              }
             }
           })
         )
@@ -62,14 +68,14 @@ export class AuthenticationService {
       map((response: any) => {
         if (response && response.token && response.expiry) {
           authContext.token = response.token;
-          sessionStorage.setItem('currentUser', JSON.stringify(authContext));
+          authContext.expiry = response.expiry;
+          sessionStorage.setItem(SESSION_AUTH_CONTEXT, JSON.stringify(authContext));
+
+          this.refreshTimeout = undefined;
 
           // emit for subject
           this.authContextEmitter.next(authContext);
-
-          // refresh token 30 seconds before expiry
-          const expiryTime = response.expiry - Date.now();
-          setTimeout(() => { this.refresh(authContext) }, expiryTime - 30000);
+          this.setupRefreshTimeout();
         }
       }),
       catchError((error: any) => {
@@ -78,8 +84,19 @@ export class AuthenticationService {
     ).subscribe();
   }
 
+  setupRefreshTimeout(): void {
+    if (this.refreshTimeout === undefined) {
+      const authContext = this.getAuthContext();
+      const expiryTime = authContext.expiry - Date.now();
+      // console.log(`access token expires in ${expiryTime - 30000} sec`);
+      this.refreshTimeout = setTimeout(() => {
+        this.refresh(authContext);
+      }, expiryTime - 30000);
+    }
+  }
+
   logout(): void {
-    sessionStorage.removeItem('currentUser');
+    sessionStorage.removeItem(SESSION_AUTH_CONTEXT);
     // emit for subject
     this.authContextEmitter.next({ securityClearance: 0 } as UserAuthContext);
   }
@@ -99,7 +116,9 @@ export class AuthenticationService {
 
   getAuthContext(): UserAuthContext {
     try {
-      const currentUser = JSON.parse(sessionStorage.getItem('currentUser'));
+      const authContext = sessionStorage.getItem(SESSION_AUTH_CONTEXT) || this.readTransplantAuthContext();
+      // don't need to manually delete the cookie as it will expire
+      const currentUser = JSON.parse(authContext);
       return <UserAuthContext>currentUser;
     }
     catch (e) {
@@ -107,8 +126,22 @@ export class AuthenticationService {
     }
   }
 
+  readTransplantAuthContext(): string {
+    const cookieAuthString = docCookies.getItem(COOKIE_AUTH_CONTEXT);
+
+    if (cookieAuthString !== null) {
+      docCookies.removeItem(COOKIE_AUTH_CONTEXT, '/', environment.BASE_DOMAIN);
+
+      // replace it
+      if (sessionStorage.getItem(SESSION_AUTH_CONTEXT) === null) {
+        sessionStorage.setItem(SESSION_AUTH_CONTEXT, cookieAuthString);
+        return cookieAuthString;
+      }
+    }
+    return null;
+  }
+
   getAuthContextStream(): Observable<UserAuthContext> {
-    this.authContextEmitter.next(this.getAuthContext());
     return this.$authContext;
   }
 }
