@@ -1,4 +1,22 @@
 export const environment = {
   production: true,
-  API_URL: 'https://ootc.makimono.me/api/v1'
+  API_URL: 'https://ootc.makimono.me/api/v1',
+  BASE_DOMAIN: 'tccc-ootc.ca',
+  languages: [
+    {
+      locale: 'en-US',
+      name: 'English',
+      url: 'https://en.tccc-ootc.ca'
+    },
+    {
+      locale: 'zh-Hant',
+      name: '中文 (繁體)',
+      url: 'https://zht.tccc-ootc.ca'
+    },
+    {
+      locale: 'zh-Hans',
+      name: '中文 (简体)',
+      url: 'https://zhs.tccc-ootc.ca'
+    }
+  ]
 };
